requirements.txt
streamlit>=1.47.0
pandas>=1.4.0
numpy>=1.23.0
matplotlib>=3.5.0
pillow>=7.1.0
tensorflow>=2.19.0
scikit-fuzzy>=0.5.0
joblib>=1.0.0
scikit-learn>=1.0.0

